snakemake-run-subrule-line-p: Bound by rule start
Otherwise, if on first subrule or the top rule line, it returns true if
there is a previous rule block whose last subrule is run.

This function still relies on the caller (here
`snakemake-indent-rule-line') to check that it is actually in a rule
block. If it is not, it will return true if below a rule block whose
last subrule is run.