mode: Unset indent-region-function
Don't use python-indent-region for indent-region-function.  At first
glance, setting indent-region-function to nil (that is, running
indent-according-to-mode on each line) seems to indent rule blocks fine,
though it will probably fail on more complex "run" values.

However, this does mean that python-indent-region is no longer used when
indent-region is called with a region that doesn't include a rule block,
so it's probably worth adding a snakemake-indent-region function that
calls python-indent-region in this case.

Re: #8