Decrease font lock precedence of command keywords
The keyword wildcard_constraints can be either a top-level command or a
rule field keyword.  The previous order in snakemake-font-lock-keywords
resulted in wildcard_constraints always be highlighted as a command key.

Rearrange the font lock list order to put the field key entry first.
This should result in the correct highlighting in every case except for
when a command key is indented (e.g., under an "if"), but this
presumably will occur much less frequently than wildcard_constraints
being used as a field keyword.