Add a terminal interface for running Snakemake
Snakemake commands are currently executed through compile.  In
general, this works fine, but it doesn't work well when Snakemake
should be executed in a different environment than the one in which
the current Emacs session was started.  For example, Guix commands
like 'guix environment ...' manipulate environmental variables to
expose particular software.  With the current setup, snakemake-program
could be set to a wrapper script that creates the environment and then
calls Snakemake:

    guix environment -l manifest.scm --ad-hoc snakemake --pure \
         -- snakemake $@

But the disadvantage of this approach is that it adds the
environmental setup time to _each_ Snakemake call.

To work better with tools like Guix, let's add an alternative
interface that allows commands to be executed in a terminal session.
Instead of the above script, snakemake-shell-file-name can be set to a
script with

    guix environment -l manifest.scm --ad-hoc snakemake --pure

Now the environmental setup cost is limited to starting the terminal.