Add a user option to always send commands to a terminal
The primary motivation behind having a terminal interface is that it
works better with tools like Guix because the environment persists
across calls.  And the primary reason one would want to use Guix with
Snakemake is so that the dependencies to a workflow/analysis can be
specified and tightly controlled.

However, the user can accidentally step outside of this controlled
environment by running a build command before running
snakemake-term-start.  The changes from this mistake can be hard to
confidently reverse and may require a wide removal of output files.

If a user wanted to make sure to only use the Guix-controlled
environment, one approach would be to remove the Snakemake executable
from the environment in which Emacs is started (most likely the
default environment), exposing it only in the workflow environments.
The problem with this is that snakemake.el requires the Snakemake
executable for things like generating lists of targets.  Also, it
would be a global change, but a user could reasonably want to control
dependencies with Guix for one project but go through the compile
interface for another.

Add a user option that allows the user to avoid the compile interface
at a global or per-project level.