graphs: Display file if Emacs lacks ImageMagick support
To display the graph, snakemake-graph and snakemake-graph-this-file
call image-mode in a buffer with the snakemake graph output.  This is
convenient because it works without using an output file, but
displaying the graph fails if Emacs was compiled without ImageMagick
support.

When Emacs doesn't have ImageMagick support, fall back to saving the
file and then displaying it.