snakemake-graph-save: Return the name of the output file
This will allow the graphing commands to save and display the graph
when the user is running an Emacs that was not built with ImageMagick
support.

When the user provides all-blank input or does not confirm an
overwrite, raise a user error so that the caller doesn't have to check
for a nil return value.