snakemake-font-lock-keywords: Increase command precedence
With the addition of 'report', the built-in names and command names
are no longer disjoint sets.  All occurrences of 'report' are
fontified as a built-in because the built-in regexp is a substring of
the command regexp.  Move the command regexp before the built-in
regexp so that the 'report:' command is highlighted as a command.