font-lock: Adjust python-font-lock-keywords handling for Emacs 27
In Emacs 27, specifically 36b64e087e (Add
'font-lock-maximum-decoration' levels for Python, 2018-07-19),
python-font-lock-keywords was changed to a list of symbols that define
different fontification levels, so we can no longer simply tack
snakemake-font-lock-keywords onto it.

Instead define a set of levels that mirrors Python mode's.  For
Snakemake mode, the same decoration is used for all levels, but at
least we will honor the level of decoration that the user has set for
Python.

Reported-by: Nicholas Knoblauch <nwknoblauch@gmail.com>

Closes #29.