Rewrite snakemake-popup as a transient
Magit switched to Transient for its popups a while go.  Magit Popup is
still around for third-party libraries that may use it, but it's not
actively developed.

At the moment, this switch is trading one external library for another
(improved) one.  However, Transient will ship with Emacs starting with
the upcoming 28.1 release.

Message-Id: <20220222030207.204401-2-kyle@kyleam.com>