mode: Add support for new module syntax
Snakemake 6.0 introduced a module syntax that looks like this:

  module other_workflow:
      snakefile:
          "other_workflow/Snakefile"

  use rule * from other_workflow as other_*

Give the new module block the same indentation and fontification as
other blocks by adding "module" to the list of words that mark a named
rule.

For the "use rule" line, highlight "use rule" as a keyword, leaving
the rest of the line to be highlighted by python.el.